+++ src/main/java/racingcar/racinggame/race/car/Car.java
@@ -1,4 +1,4 @@
-package racingcar.racinggame.car;
+package racingcar.racinggame.race.car;
 
 import java.util.List;
